Refactors wfs-layer
