+++ src/signals/incident/components/form/ContainerSelect/types.ts
@@ -1,5 +1,4 @@
-import type { IconOptions, LatLngExpression, Map } from 'leaflet';
-import type { ZoomLevel } from '@amsterdam/arm-core/lib/types';
+import type { IconOptions, LatLngExpression } from 'leaflet';
 
 export type ClickEvent = (e: React.MouseEvent<HTMLButtonElement | HTMLAnchorElement>) => void;
 
@@ -42,14 +41,6 @@ export interface ContainerSelectValue {
   update: (items: Item[] | null) => void;
   edit: () => void;
   close: () => void;
-}
-
-export interface WfsLayerProps {
-  url: string;
-  options: {
-    getBbox: (mapInstance: Map) => string;
-  };
-  zoomLevel: ZoomLevel;
 }
 
 export interface DataLayerProps {
